refactor(library/Attempt): moves "Adapted/asynchronous.ts" to "Adapted/toSettle.ts"
