+++ src/library/Attempt/Adapted/toSettle.ts
